requirements.txt
requests~=2.32.5
serpapi~=0.1.5
Flask~=3.1.2
Flask-Login~=0.6.3
cryptography~=46.0.1
geopy~=2.4.1
click~=8.3.0
python-dotenv~=1.1.1
oauthlib~=3.3.1